requirements.txt
pyrogram>=2.0.0
yt-dlp>=2023.12.30
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio
